feat(@cubejs-backend/server): Migrate CubejsServer to TypeScript
