--- packages/cubejs-server/index.d.ts
@@ -1,19 +1,0 @@
-import { CreateOptions as CoreCreateOptions } from "@cubejs-backend/server-core";
-import express from 'express';
-import https from 'https';
-import http from 'http';
-
-export interface CreateOptions extends CoreCreateOptions {
-  webSockets?: boolean;
-  initApp?(app: express.Application): void | Promise<void>;
-}
-
-declare class CubejsServer {
-  constructor(options: CreateOptions);
-
-  listen(options?: https.ServerOptions | http.ServerOptions): Promise<{ server: http.Server | https.Server, version: string, port: number, tlsPort?: number, app: express.Application }>;
-  close(): Promise<void>;
-  testConnections(): Promise<void>;
-}
-
-export default CubejsServer;
