+++ packages/cubejs-server/package.json
@@ -10,10 +10,10 @@
     "directory": "packages/cubejs-server"
   },
   "engines": {
-    "node": ">=8.11.1"
+    "node": ">=10.0"
   },
   "main": "src/src/index.js",
-  "typings": "index.d.ts",
+  "typings": "src/dist/index.d.ts",
   "files": [
     "dist/src",
     "index.d.ts",
@@ -47,6 +47,8 @@
     "ws": "^7.1.2"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.8",
+    "@types/node": "^10.0.10",
     "eslint": "^6.8.0",
     "eslint-plugin-node": "^5.2.1",
     "jest": "^24.9.0",
